Add endpoints: Images, Models, Moderation.

+++ src/main/java/org/haifan/merlin/model/openai/models/Model.java
@@ -9,24 +9,12 @@ import lombok.Data;
  */
 @Data
 public class Model {
-    /**
-     * The model identifier, which can be referenced in the API endpoints.
-     */
     private String id;
 
-    /**
-     * The type of object returned, should be "model"
-     */
     private String object;
 
-    /**
-     * The Unix timestamp (in seconds) when the model was created.
-     */
-    private Integer created;
+    private Long created;
 
-    /**
-     * The organization that owns the model.
-     */
     @JsonProperty("owned_by")
     private String ownedBy;
 }
